Pencil Mark Progress
Changed wording to penciling instead of solving, as solving will use the pencil grid, but go through a separate loop/process. Pencil marking is being separated into a different module.

Instead of adding the values and then cycling through to invert for pencil marking, we are going to just start with a full 1-9 in every cell and just remove as we go. Makes more sense.

In the process of shifting pencil mark functions over.