Models and JSDoc
Introduced a model for the Grid object.

The reason I want to switch to TypeScript in the future is to ensure consistency in how I read my code and how it performs. This is a good step towards both of those things.

Cleaned up some comments and debugging statements. Installed JSDoc in preparation for better documentation/readability.