TypeScript Migration
A few random issues were arising from not being in TS, namely losing track of data flow and inferring types, so I decided to start migrating to TS.

Installed TS packages and Added interfaces for the three grids (two interfaces are redundant, but still included) and the selected cell. Added classes for each type of grid as well that implement the interfaces.

Finished TS implementation on the Solver component.