Fixed penciling
Added property for manual filling of pencil marks in order to properly populate the grid. Changed names to reflect this being a manual process.

Fixed the issue with the value variable adding 1 as a string, causing the functions after to never touch 0-8.

Disabled and styled the solve button to indicate it is unfinished.